requirements.txt
streamlit
pandas
plotly
openpyxl
gspread
google-auth
google-api-python-client
yagmail